+++ .github/workflows/check-trigger.yml
@@ -14,5 +14,8 @@ jobs:
       - uses: actions/setup-python@v4
         with:
           python-version: '3.11'
-      - name: check trigger
+          cache: 'pip'
+      - name: Install Dependencies
+        run: python -m pip install -r requirements.txt
+      - name: Run Check Trigger Script
         run: python src/check_trigger.py
