+++ requirements.txt
@@ -2,4 +2,5 @@
 jupyter==1.0.0
 jupytext==1.11.1
 jupyter-black==0.3.1
+pandas==2.0.3
 pre-commit==2.15.0
