+++ src/utils.py
@@ -5,7 +5,6 @@ import re
 import xml.etree.ElementTree as ET
 import zipfile
 from datetime import datetime
-from io import StringIO
 from pathlib import Path
 from typing import Dict, List
 
@@ -18,6 +17,8 @@ from dotenv import load_dotenv
 from ochanticipy.utils.hdx_api import load_resource_from_hdx
 from shapely.geometry import LineString
 from tqdm.auto import tqdm
+
+from src.check_trigger import datetime_to_season, load_fms_forecast
 
 load_dotenv()
 
@@ -49,51 +50,6 @@ MB_TOKEN = os.environ.get("MAPBOX_TOKEN")
 CS_KEY = os.environ.get("CHARTSTUDIO_APIKEY")
 MAPS_DIR = PROC_PATH / "historical_forecast_maps"
 STORM_DATA_DIR = Path(os.getenv("STORM_DATA_DIR"))
-
-
-def load_fms_forecast(path: Path | StringIO) -> pd.DataFrame:
-    """
-    Loads FMS raw forecast
-    Parameters
-    ----------
-    path: Path | StringIO
-        Path to raw forecast CSV. Path can be a StringIO
-        (so CSV can be passed as an encoded string from Power Automate)
-
-
-    Returns
-    -------
-    DataFrame of processed forecast
-    """
-    df_date = pd.read_csv(path, header=None, nrows=3)
-    date_str = df_date.iloc[0, 1].removeprefix("baseTime=")
-    base_time = datetime.strptime(date_str, "%Y-%m-%dT%H:%M:%SZ")
-    cyclone_name = (
-        df_date.iloc[2, 0].removeprefix("# CycloneName=").capitalize()
-    )
-    if isinstance(path, StringIO):
-        path.seek(0)
-    df_data = pd.read_csv(
-        path,
-        skiprows=range(6),
-    )
-    df_data = df_data.drop([0])
-    df_data = df_data.rename(
-        columns={"Time[fmt=yyyy-MM-dd'T'HH:mm:ss'Z']": "forecast_time"}
-    )
-    df_data["forecast_time"] = pd.to_datetime(
-        df_data["forecast_time"]
-    ).dt.tz_localize(None)
-    df_data["cyclone_name"] = cyclone_name
-    df_data["base_time"] = base_time
-    df_data["season"] = datetime_to_season(base_time)
-    df_data["Name Season"] = df_data["cyclone_name"] + " " + df_data["season"]
-    df_data["leadtime"] = df_data["forecast_time"] - df_data["base_time"]
-    df_data["leadtime"] = (
-        df_data["leadtime"].dt.days * 24
-        + df_data["leadtime"].dt.seconds / 3600
-    ).astype(int)
-    return df_data
 
 
 def check_fms_forecast(
@@ -1019,9 +975,3 @@ def load_buffer(distance: int = 250) -> gpd.GeoDataFrame:
     load_path = PROC_PATH / "buffer" / filename / f"{filename}.shp"
 
     return gpd.read_file(load_path)
-
-
-def datetime_to_season(date):
-    # July 1 (182nd day of the year) is technically the start of the season
-    eff_date = date - pd.Timedelta(days=182)
-    return f"{eff_date.year}/{eff_date.year + 1}"
